例外発生時の処理を見直し。run_instanceの途中でterminateされてしまう

+++ lib/vagrant-niftycloud/action/terminate_instance.rb
@@ -17,46 +17,46 @@ module VagrantPlugins
           # http://cloud.nifty.com/api/sdk/rdoc/
           begin
             env[:ui].info(I18n.t("vagrant_niftycloud.terminating"))
+	    
+	    # 起動直後等、terminate処理できないステータスの場合一旦待つ
+            server = env[:niftycloud_compute].describe_instances(:instance_id => env[:machine].id).reservationSet.item.first.instancesSet.item.first
+            while server.instanceState.name == 'pending'
+              sleep 5
+              server = env[:niftycloud_compute].describe_instances(:instance_id => env[:machine].id).reservationSet.item.first.instancesSet.item.first
+            end
 
             attribute = env[:niftycloud_compute].describe_instance_attribute(:instance_id => env[:machine].id, :attribute => 'disableApiTermination')
-            if attribute.disableApiTermination.value != 'false'
-              # API経由で立ち上げていない場合、即terminateができないため念のため一旦stopする
+            if attribute.disableApiTermination.value == 'false'
+              # AWSのように即terminateができないため念のため一旦stopする
+              # TODO API経由でのterminate不可の場合を考慮する必要があるか
               server = env[:niftycloud_compute].describe_instances(:instance_id => env[:machine].id).reservationSet.item.first.instancesSet.item.first
               if server.instanceState.name != 'stopped'
                 env[:niftycloud_compute].stop_instances(:instance_id => env[:machine].id, :force => true)
                 while server.instanceState.name != 'stopped'
+                  sleep 5
                   server = env[:niftycloud_compute].describe_instances(:instance_id => env[:machine].id).reservationSet.item.first.instancesSet.item.first
-                  sleep 5
-                end
-              end
-
-              if config[:force] == false
-                # TODO APIからの削除拒否の場合。考慮する必要はあるか
-              else
-                env[:niftycloud_compute].modify_instance_attribute(:instance_id => env[:machine].id, :attribute => 'disableApiTermination', :value => 'false')
-                while attribute.disableApiTermination.value != 'false'
-                  attribute = env[:niftycloud_compute].describe_instance_attribute(:instance_id => env[:machine].id, :attribute => 'disableApiTermination')
-                  sleep 5
                 end
               end
             end
 
-            #terminate処理
-            response = env[:niftycloud_compute].terminate_instances(:instance_id => env[:machine].id)
-            env[:machine].id = nil
+            # terminate処理
+            server = env[:niftycloud_compute].describe_instances(:instance_id => env[:machine].id).reservationSet.item.first.instancesSet.item.first
+            if server.instanceState.name == 'stopped'
+              response = env[:niftycloud_compute].terminate_instances(:instance_id => env[:machine].id)
+              env[:machine].id = nil
 
-            @app.call(env)
-          rescue ConfigurationError => e
-            raise Errors::NiftyCloudConfigurationError,
+              @app.call(env)
+	    end
+          rescue NIFTY::ConfigurationError => e
+            raise VagrantPlugins::NiftyCloud::Errors::NiftyCloudConfigurationError,
               :code    => e.error_code,
               :message => e.error_message
-          rescue ArgumentError => e
-            raise Errors::NiftyCloudArgumentError,
-              :code    => e.error_code,
-              :message => e.error_message
-          rescue => e
             ui.error("Could not locate server '#{env[:machine].id}'.  Please verify it was provisioned in the current zone.")
-          end
+          rescue NIFTY::ArgumentError => e
+            raise VagrantPlugins::NiftyCloud::Errors::NiftyCloudArgumentError,
+              :code    => e.error_code,
+              :message => e.error_message
+	  end
         end
       end
     end
